+++ examples/plugin_config.py
@@ -1,4 +1,7 @@
-# XRLint configuration file example
+"""
+This configuration example shows how to define and use a plugin
+using the `Plugin` class and its `define_rule()` decorator method.
+"""
 
 from xrlint.config import Config
 from xrlint.node import DatasetNode
@@ -11,21 +14,24 @@ from xrlint.rule import RuleOp
 plugin = Plugin(
     meta=PluginMeta(name="hello-plugin", version="1.0.0"),
     configs={
+        # "configs" entries must be `Config` objects!
         "recommended": Config.from_value(
             {
                 "rules": {
-                    "hello/good-title": "error",
+                    "hello/good-title": "warn",
+                    # Configure more rules here...
                 },
             }
-        )
+        ),
+        # Add more configurations here...
     },
 )
 
 
-@plugin.define_rule(
-    "good-title", description=f"Dataset title should be 'Hello World!'."
-)
+@plugin.define_rule("good-title")
 class GoodTitle(RuleOp):
+    """Dataset title should be 'Hello World!'."""
+
     def dataset(self, ctx: RuleContext, node: DatasetNode):
         good_title = "Hello World!"
         if node.dataset.attrs.get("title") != good_title:
@@ -35,16 +41,19 @@ class GoodTitle(RuleOp):
             )
 
 
+# Define more rules here...
+
+
 def export_configs():
     return [
-        {
-            "files": ["**/*.zarr", "**/*.nc"],
-        },
+        # Use "hello" plugin
         {
             "plugins": {
                 "hello": plugin,
             },
         },
+        # Use recommended settings from xrlint
         "recommended",
+        # Use recommended settings from "hello" plugin
         "hello/recommended",
     ]