+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Version 0.2.0 (in development)
 
 - Make all docstrings comply to google-style
+- Rule description is now your `RuleOp`'s docstring
+  if `description` is not explicitly provided.
 
 ## Version 0.1.0 (09.01.2025)